refactor: security 리팩토링
