--- api-module/src/main/java/com/foodgo/apimodule/user/exception/UserExceptionHandler.java
@@ -1,10 +1,0 @@
-package com.foodgo.apimodule.user.exception;
-
-import com.foodgo.commonmodule.exception.common.BaseErrorCode;
-import com.foodgo.commonmodule.exception.common.CustomException;
-
-public class UserExceptionHandler extends CustomException {
-	public UserExceptionHandler(BaseErrorCode code) {
-		super(code);
-	}
-}
