+++ api-module/src/main/resources/application-prod.yml
@@ -27,8 +27,8 @@ spring:
 jwt:
   secret: aGFuZXVtLWZvb2Rnby1qd3Qtc2VjcmV0LWtleQo=
   token:
-    access-expiration-time: 86400
-    refresh-expiration-time: 604800
+    access-expiration-time: 86400000 # 24 * 60 * 60 * 1000 = 1일
+    refresh-expiration-time: 604800000 # 7 * 24 * 60 * 60 * 1000 = 7일
 
 logging:
   level:
