+++ common-module/src/main/java/com/foodgo/commonmodule/common/BaseErrorCode.java
@@ -1,0 +1,14 @@
+package com.foodgo.commonmodule.common;
+
+import org.springframework.http.HttpStatus;
+
+public interface BaseErrorCode {
+
+    HttpStatus getHttpStatus();
+
+    String getCode();
+
+    String getMessage();
+
+    ApiResponse<Void> getErrorResponse();
+}
