+++ common-module/src/main/java/com/foodgo/commonmodule/common/CustomException.java
@@ -1,0 +1,13 @@
+package com.foodgo.commonmodule.common;
+
+import lombok.Getter;
+
+@Getter
+public class CustomException extends RuntimeException {
+
+	private final BaseErrorCode errorCode;
+
+	public CustomException(BaseErrorCode errorCode) {
+		this.errorCode = errorCode;
+	}
+}
