+++ common-module/src/main/java/com/foodgo/commonmodule/common/TempHandler.java
@@ -1,0 +1,8 @@
+package com.foodgo.commonmodule.common;
+
+public class TempHandler extends CustomException {
+
+	public TempHandler(BaseErrorCode errorCode) {
+		super(errorCode);
+	}
+}
