--- common-module/src/main/java/com/foodgo/commonmodule/exception/common/BaseErrorCode.java
@@ -1,14 +1,0 @@
-package com.foodgo.commonmodule.exception.common;
-
-import org.springframework.http.HttpStatus;
-
-public interface BaseErrorCode {
-
-    HttpStatus getHttpStatus();
-
-    String getCode();
-
-    String getMessage();
-
-    ApiResponse<Void> getErrorResponse();
-}
