--- common-module/src/main/java/com/foodgo/commonmodule/exception/common/CustomException.java
@@ -1,13 +1,0 @@
-package com.foodgo.commonmodule.exception.common;
-
-import lombok.Getter;
-
-@Getter
-public class CustomException extends RuntimeException {
-
-	private final BaseErrorCode errorCode;
-
-	public CustomException(BaseErrorCode errorCode) {
-		this.errorCode = errorCode;
-	}
-}
