--- common-module/src/main/java/com/foodgo/commonmodule/exception/common/TempHandler.java
@@ -1,8 +1,0 @@
-package com.foodgo.commonmodule.exception.common;
-
-public class TempHandler extends CustomException {
-
-	public TempHandler(BaseErrorCode errorCode) {
-		super(errorCode);
-	}
-}
