--- common-module/src/main/java/com/foodgo/commonmodule/exception/jwt/dto/JwtDto.java
@@ -1,7 +1,0 @@
-package com.foodgo.commonmodule.exception.jwt.dto;
-
-public record JwtDto(
-	String accessToken,
-	String refreshToken
-) {
-}
