+++ common-module/src/main/java/com/foodgo/commonmodule/jwt/dto/JwtDto.java
@@ -1,0 +1,7 @@
+package com.foodgo.commonmodule.jwt.dto;
+
+public record JwtDto(
+	String accessToken,
+	String refreshToken
+) {
+}
