+++ common-module/src/main/java/com/foodgo/commonmodule/security/config/CorsConfig.java
@@ -20,6 +20,7 @@ public class CorsConfig implements WebMvcConfigurer {
         ArrayList<String> allowedOriginPatterns = new ArrayList<>();
         allowedOriginPatterns.add("http://localhost:8080");
         allowedOriginPatterns.add("http://localhost:3000");
+        allowedOriginPatterns.add("https://foodgohaneum.link");
 
         ArrayList<String> allowedHttpMethods = new ArrayList<>();
         allowedHttpMethods.add("GET");
