+++ common-module/src/main/java/com/foodgo/commonmodule/user/exception/UserExceptionHandler.java
@@ -1,0 +1,10 @@
+package com.foodgo.commonmodule.user.exception;
+
+import com.foodgo.commonmodule.common.BaseErrorCode;
+import com.foodgo.commonmodule.common.CustomException;
+
+public class UserExceptionHandler extends CustomException {
+	public UserExceptionHandler(BaseErrorCode code) {
+		super(code);
+	}
+}
