+++ core-module/src/main/java/com/foodgo/coremodule/user/application/UserQueryService.java
@@ -3,12 +3,13 @@ package com.foodgo.coremodule.user.application;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
-import com.foodgo.apimodule.user.exception.UserErrorCode;
-import com.foodgo.apimodule.user.exception.UserExceptionHandler;
+import com.foodgo.commonmodule.user.exception.UserErrorCode;
+import com.foodgo.commonmodule.user.exception.UserExceptionHandler;
 import com.foodgo.coremodule.user.domain.User;
 import com.foodgo.coremodule.user.repository.UserRepository;
 
 import lombok.RequiredArgsConstructor;
+
 @RequiredArgsConstructor
 @Transactional(readOnly = true)
 @Service
