--- core-module/src/main/java/com/foodgo/coremodule/user/repository/UserJpaRepository.java
@@ -1,12 +1,0 @@
-package com.foodgo.coremodule.user.repository;
-
-import java.util.Optional;
-
-import org.springframework.data.jpa.repository.JpaRepository;
-
-import com.foodgo.coremodule.user.domain.User;
-
-public interface UserJpaRepository extends JpaRepository<User, Long> {
-
-	Optional<User> findByUsername(String username);
-}
