+++ core-module/src/main/java/com/foodgo/coremodule/user/repository/UserRepository.java
@@ -2,9 +2,11 @@ package com.foodgo.coremodule.user.repository;
 
 import java.util.Optional;
 
+import org.springframework.data.jpa.repository.JpaRepository;
+
 import com.foodgo.coremodule.user.domain.User;
 
-public interface UserRepository {
+public interface UserRepository extends JpaRepository<User, Long>, UserRepositoryCustom {
 
 	Optional<User> findByUsername(String username);
