+++ core-module/src/main/java/com/foodgo/coremodule/user/repository/UserRepositoryCustom.java
@@ -1,0 +1,4 @@
+package com.foodgo.coremodule.user.repository;
+
+public interface UserRepositoryCustom {
+}
